requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
Pillow==10.1.0
requests==2.31.0
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
